class work, search pipe, got ride of tv and book components b/c I want to do those differently
