+++ src/app/app.module.ts
@@ -1,12 +1,17 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { AppComponent } from './app.component';
 import { NavbarComponent } from './navbar/navbar.component';
 import { HomepageComponent } from './homepage/homepage.component';
 import { RoutingModule } from './app.routes';
-import { MoviesComponent } from './movies/movies.component';
-import { TelevisionComponent } from './television/television.component';
-import { BooksComponent } from './books/books.component';
+import { MovieListComponent } from './movie-list/movie-list.component';
+import { ItemDetailComponent } from './item-detail/item-detail.component';
+import { MovieService } from './services/movie.service';
+
+//Custom
+import { SearchByPipe } from './pipes/search-by.pipe';
+
 
 
 @NgModule({
@@ -14,17 +19,18 @@ import { BooksComponent } from './books/books.component';
     AppComponent,
     NavbarComponent,
     HomepageComponent,
-    MoviesComponent,
-    TelevisionComponent,
-    BooksComponent,
+    MovieListComponent,
+    ItemDetailComponent,
+    SearchByPipe,
 
   ],
   imports: [
     BrowserModule,
+    FormsModule,
     //ROUTING MODULE SHOULD BE LAST
     RoutingModule
   ],
-  providers: [],
+  providers: [MovieService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
