--- src/app/books/books.component.html
@@ -1,3 +1,0 @@
-<p>
-  books works!
-</p>
