--- src/app/books/books.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-books',
-  templateUrl: './books.component.html',
-  styleUrls: ['./books.component.css']
-})
-export class BooksComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
