--- src/app/homepage/homepage.component.html
@@ -1,12 +1,0 @@
-<div *ngFor="let movie of movies" class="movies">
-  <ul>
-    <li>
-      <h2>{{ movie.title }}</h2>
-      <img src="{{ movie.thumbnail }}" alt="">
-      <p>{{ movie.released_on }}</p>
-      <p>{{ movie.genre }}</p>
-      <p>{{ movie.production_companies }}</p>
-      <p>{{ movie.description }}</p>
-    </li>
-  </ul>
-</div>
