+++ src/app/item-detail/item-detail.component.html
@@ -1,0 +1,3 @@
+<p>
+  item-detail works!
+</p>
