+++ src/app/models/movie.ts
@@ -1,0 +1,14 @@
+export class Movie {
+  title: string;
+  released_on: string;
+  genre: any;
+  description: any;
+  actors: any;
+  production_companies: any;
+  run_time: string;
+  rating: string;
+  director: any;
+  writing_credits: any;
+  thumbnail: string;
+  id: string;
+}
