+++ src/app/models/movies.ts
@@ -1,0 +1,894 @@
+import { Movie } from './movie';
+
+export const Movies: Array<Movie> = [
+    {
+     title: 'Howl\'s Moving Castle',
+     released_on: '2004',
+     genre: [
+       'Animation',
+       'Adventure',
+       'Family',
+     ],
+     description: 'When an unconfident young woman is cursed with an old body by a spiteful witch, her only chance of breaking the spell lies with a self-indulgent yet insecure young wizard and his companions in his legged, walking castle.',
+     actors: [
+       'Christian Bale',
+       'Jean Simmons',
+       'Lauren Bacall',
+       'Blythe Danner',
+       'Billy Crystal',
+       'Emily Mortimer',
+       'Chieko Baishô',
+       'Takuya Kimura',
+       'Akihiro Miwa',
+       'Tatsuya Gashûin',
+       'Ryûnosuke Kamiki',
+     ],
+     production_companies: [
+       'Studio Ghibli',
+       'Tokuma Shoten',
+       'd-rights',
+       'Nippon Television Network',
+       'Dentsu',
+       'Mitsubishi',
+       'Tohokushinsha Film Corporation',
+       'Buena Vista Home Entertainment',
+       'DENTSU Music and Entertainment',
+     ],
+     run_time: '1 Hour 59 Minutes',
+     rating: 'PG',
+     director: 'Hiyao Miyazaki',
+     writing_credits: [
+       'Hiyao Miyazaki',
+       'Dianna Wynne Jones',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BZTRhY2QwM2UtNWRlNy00ZWQwLTg3MjktZThmNjQ3NTdjN2IxXkEyXkFqcGdeQXVyMzg2MzE2OTE@._V1_UY268_CR5,0,182,268_AL_.jpg',
+     id: '1',
+   },
+   {
+     title: 'She\'s The Man',
+     released_on: '2006',
+     genre: [
+       'Comedy',
+       'Romance',
+       'Sport',
+     ],
+     description: 'When her brother decides to ditch for a couple weeks, Viola heads over to his elite boarding school, disguised as him, and proceeds to fall for one of his soccer teammates, and soon learns she\'s not the only one with romantic troubles.',
+     actors: [
+       'Amanda Bynes',
+       'Channing Tatum',
+       'Laura Ramsey',
+       'Vinnie Jones',
+       'Robert Hoffman',
+       'Alex Breckenridge',
+       'Julie Hagerty',
+       'David Cross',
+     ],
+     production_companies: [
+       'DreamWorks',
+       'Lakeshore Entertainment',
+       'Donners\' Company',
+     ],
+     run_time: '1 Hour 45 Minutes',
+     rating: 'PG-13',
+     director: 'Andy Fickman',
+     writing_credits: [
+       'Ewan Leslie',
+       'Karen McCullah',
+       'Kirsten Smith',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjEyMzAzMDk1Ml5BMl5BanBnXkFtZTcwNjg0OTEzMQ@@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '20',
+   },
+   {
+     title: 'How to Deal',
+     released_on: '2003',
+     genre: [
+       'Comedy',
+       'Drama',
+       'Romance',
+     ],
+     description: 'For teenager Halley, love seems like an alien concept. Her mother, Lydia, and father, Len, are divorced. Lydia doesn\'t date, and Len is now seeing a younger woman. Halley\'s sister, Ashley, is about to get married and doesn\'t realize that her fiancé is not the man of her dreams. But, when Halley meets edgy Macon, she learns what love is really all about.',
+     actors: [
+       'Mandy Moore',
+       'Allison Janney',
+       'Trent Ford',
+       'Alexandra Holden',
+       'Dylan Baker',
+       'Peter Gallagher',
+     ],
+     production_companies: [
+       'New Line Cinema',
+       'Golden Mean',
+       'Radar Pictures',
+     ],
+     run_time: '1 Hour 41 Minutes',
+     rating: 'PG-13',
+     director: 'Claire Kilner',
+     writing_credits: [
+       'Sarah Dessen',
+       'Neena Beber',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BOTMzOTk0NDA4OV5BMl5BanBnXkFtZTYwNTY5NjY3._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '2',
+   },
+   {
+     title: 'Digimon: The Movie',
+     released_on: '2000',
+     genre: [
+       'Animation',
+       'Action',
+       'Adventure',
+     ],
+     description: 'Three different stories involving digital creatures collide together: 1. The Digidestined-kids chosen to save the digital world-come face to face with their first Digimon. 2. When a powerful new internet Digimon hatches and begins to consume data at an alarming rate, the Digidestined must put an end to the destruction before the damage becomes irreversable and worldwide communication halts forever. 3. As computer-based missiles are launched, and a wayward Digimon kidnaps the Digidestined, only the combined efforts of a worldwide network of kids and a new group of "Digidestined" can rescue the others and stop global disaster.',
+     actors: [
+       'Lara Jill Miller',
+       'Joshua Seth',
+       'Bob Papenbrook',
+       'Jeff Nimoy',
+       'Colleen O\'Shaughnessey',
+       'Philece Sampler',
+       'Mona Marshall',
+       'Michael Lindsay',
+       'Michael Reisz',
+       'Wendee Lee',
+       'Doug Erholtz',
+       'Bob Glouberman',
+       'Brian Donovan',
+     ],
+     production_companies: [
+       'Saban Entertainment',
+       'Toei Animation',
+       'Toei Company',
+       'Fox Kids Network',
+       'Fox Family Films',
+       'Fox Kids Europe Properties',
+       'Fox Kids International Programming',
+     ],
+     run_time: '1 Hour 22 Minutes',
+     rating: 'PG',
+     director: [
+       'Mamoru Hosoda',
+       'Shigeyasu Yamauchi',
+     ],
+     writing_credits: [
+       'Akiyoshi Hongo',
+       'Jeff Nimoy',
+       'Bob Buchholz',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjE2NTAyMjUxMl5BMl5BanBnXkFtZTYwOTAwODQ3._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '3',
+   },
+   {
+     title: 'Saved',
+     released_on: '2004',
+     genre: [
+       'Comedy',
+       'Drama',
+     ],
+     description: 'When a girl attending a Christian high school becomes pregnant, she finds herself ostracized and demonized, as all of her former friends turn on her.',
+     actors: [
+       'Jena Malone',
+       'Mandy Moore',
+       'Patrick Fugit',
+       'Macaulay Culkin',
+       'Eva Amurri Martino',
+       'Mary-Louise Parker',
+     ],
+     production_companies: [
+       'United Artists',
+       'Single Cell Pictures',
+       'Infinity Media',
+       'James Forsyth Casting',
+       'Red Bull Productions',
+     ],
+     run_time: '1 Hour 32 Minutes',
+     rating: 'PG-13',
+     director: 'Brian Dannelly',
+     writing_credits: [
+       'Brian Dannelly',
+       'Michael Urban',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMTQ1ODM1ODU1N15BMl5BanBnXkFtZTcwMjI5MzA3NA@@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '4',
+   },
+   {
+     title: 'Kill Bill: Volume 1',
+     released_on: '2003',
+     genre: [
+       'Action',
+       'Crime',
+       'Thriller',
+     ],
+     description: 'The Bride wakens from a four-year coma. The child she carried in her womb is gone. Now she must wreak vengeance on the team of assassins who betrayed her - a team she was once part of.',
+     actors: [
+       'Uma Thurman',
+       'David Carradine',
+       'Daryl Hannah',
+       'Lucy Liu',
+       'Vivica A. Fox'
+     ],
+     production_companies: [
+       'Miramax',
+       'A Band Apart',
+       'Super Cool Manchu',
+     ],
+     run_time: '1 Hour 51 Minutes',
+     rating: 'R',
+     director: 'Quentin Tarantino',
+     writing_credits: 'Quentin Tarantino',
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BYTczMGFiOWItMjA3Mi00YTU5LWIwMDgtYTEzNjRkNDkwMTE2XkEyXkFqcGdeQXVyNzQ1ODk3MTQ@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '5',
+   },
+   {
+     title: 'Kill Bill: Volume 2',
+     released_on: '2004',
+     genre: [
+       'Action',
+       'Crime',
+       'Thriller',
+     ],
+     description: 'The Bride continues her quest of vengeance against her former boss and lover Bill, the reclusive bouncer Budd and the treacherous, one-eyed Elle.',
+     actors: [
+       'Uma Thurman',
+       'David Carradine',
+       'Michael Madsen',
+       'Michael Bowen',
+       'Lucy Liu',
+       'Vivica A. Fox',
+       'Samuel L. Jackson',
+     ],
+     production_companies: [
+       'Miramax',
+       'A Band Apart',
+       'Super Cool Manchu',
+     ],
+     run_time: '2 Hours 17 Minutes',
+     rating: 'R',
+     director: 'Quentin Tarantino',
+     writing_credits: 'Quentin Tarantino',
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BNmFiYmJmN2QtNWQwMi00MzliLThiOWMtZjQxNGRhZTQ1MjgyXkEyXkFqcGdeQXVyNzQ1ODk3MTQ@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '6',
+   },
+   {
+     title: 'Pride & Prejudice',
+     released_on: '2005',
+     genre: [
+       'Drama',
+       'Romance',
+     ],
+     description: 'Sparks fly when spirited Elizabeth Bennet meets single, rich, and proud Mr. Darcy. But Mr. Darcy reluctantly finds himself falling in love with a woman beneath his class. Can each overcome their own pride and prejudice?',
+     actors: [
+       'Keira Knightly',
+       'Matthew Macfayden',
+       'Brenda Blethyn',
+       'Donald Sutherland',
+       'Rosamund Pike',
+       'Jena Malone',
+       'Talulah Riley',
+       'Carey Mulligan',
+       'Claudie Blakley',
+       'Simon Woods',
+       'Rupert Friend',
+       'Judi Dench',
+     ],
+     production_companies: [
+       'Focus Features',
+       'Universal Pictures',
+       'Studio Canal',
+       'Working Title Films',
+       'Scion Films',
+     ],
+     run_time: '2 Hours 9 Minutes',
+     rating: 'PG',
+     director: 'Joe Wright',
+     writing_credits: [
+       'Jane Austen',
+       'Deborah Moggach',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMTA1NDQ3NTcyOTNeQTJeQWpwZ15BbWU3MDA0MzA4MzE@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '7',
+   },
+   {
+     title: 'Atomic Blonde',
+     released_on: '2017',
+     genre: [
+       'Action',
+       'Mystery',
+       'Thriller',
+     ],
+     description: 'An undercover MI6 agent is sent to Berlin during the Cold War to investigate the murder of a fellow agent and recover a missing list of double agents.',
+     actors: [
+       'Charlize Theron',
+       'James McAvoy',
+       'Eddie Marsan',
+       'John Goodman',
+       'Toby Jones',
+       'Roland Møller',
+       'Sofia Boutella',
+       'Bill Skarsgård',
+     ],
+     production_companies: [
+       '87Eleven',
+       'Closed on Mondays Entertainment',
+       'Denver and Delilah Productions',
+       'Film i Väst',
+       'Sierra / Affinity',
+       'T.G.I.M. Films',
+     ],
+     run_time: '1 Hour 55 Minutes',
+     rating: 'R',
+     director: 'David Leitch',
+     writing_credits: [
+       'Kurt Johnstad',
+       'Antony Johnston',
+       'Sam Hart',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjM5NDYzMzg5N15BMl5BanBnXkFtZTgwOTM2NDU1MjI@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '8',
+   },
+   {
+     title: 'The Princess Bride',
+     released_on: '1987',
+     genre: [
+       'Adventure',
+       'Family',
+       'Fantasy',
+     ],
+     description: 'While home sick in bed, a young boy\'s grandfather reads him a story called The Princess Bride.',
+     actors: [
+       'Robin Wright',
+       'Cary Elwes',
+       'Mandy Patinkin',
+       'André the Giant',
+       'Chris Sarandon',
+       'Christopher Guest',
+       'Wallace Shawn',
+       'Fred Savage',
+       'Peter Falk',
+     ],
+     production_companies: [
+       'Act III Communications',
+       'Buttercup Films Ltd.',
+       'The Princess Bride Ltd',
+     ],
+     run_time: '1 Hour 38 Minutes',
+     rating: 'PG',
+     director: 'Rob Reiner',
+     writing_credits: 'William Goldman',
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMGM4M2Q5N2MtNThkZS00NTc1LTk1NTItNWEyZjJjNDRmNDk5XkEyXkFqcGdeQXVyMjA0MDQ0Mjc@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '9',
+   },
+   {
+     title: 'Guardians of the Galaxy',
+     released_on: '2014',
+     genre: [
+       'Action',
+       'Adventure',
+       'Sci-Fi',
+     ],
+     description: 'A group of intergalactic criminals are forced to work together to stop a fanatical warrior from taking control of the universe.',
+     actors: [
+       'Chris Pratt',
+       'Zoe Saldana',
+       'Dave Bautista',
+       'Vin Diesel',
+       'Bradley Cooper',
+       'Lee Pace',
+       'Michael Rooker',
+       'Karen Gillan',
+       'John C. Reilly',
+       'Glenn Close',
+       'Djimon Hounsou',
+       'Benicio Del Toro',
+       'Sean Gunn',
+       'Peter Serafinowicz',
+     ],
+     production_companies: [
+       'Marvel Studios',
+       'Marvel Enterprises',
+       'Moving Picture Company',
+     ],
+     run_time: '2 Hours 1 Minute',
+     rating: 'PG-13',
+     director: 'James Gunn',
+     writing_credits: [
+       'James Gunn',
+       'Nicole Perlman',
+       'Dan Abnett',
+       'Andy Lanning',
+       'Bill Mantlo',
+       'Keith Giffen',
+       'Jim Starlin',
+       'Steve Englehart',
+       'Steve Gan',
+       'Steve Gerber',
+       'Val Mayerik',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMTAwMjU5OTgxNjZeQTJeQWpwZ15BbWU4MDUxNDYxODEx._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '10',
+   },
+   {
+     title: 'Guardians of the Galaxy Vol. 2',
+     released_on: '2017',
+     genre: [
+       'Action',
+       'Adventure',
+       'Sci-Fi',
+     ],
+     description: 'The Guardians must fight to keep their newfound family together as they unravel the mystery of Peter Quill\'s true parentage.',
+     actors: '',
+     production_companies: [
+       'Marvel Studios',
+       'Walt Disney Pictures',
+     ],
+     run_time: '2 Hours 16 Minutes',
+     rating: 'PG-13',
+     director: 'James Gunn',
+     writing_credits: [
+       'James Gunn',
+       'Dan Abnett',
+       'Andy Lanning',
+       'Steve Englehart',
+       'Steve Gan',
+       'Jim Starlin',
+       'Stan Lee',
+       'Larry Lieber',
+       'Jack Kirby',
+       'Bill Mantlo',
+       'Keith Giffen',
+       'Steve Gerber',
+       'Val Mayerik',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMTg2MzI1MTg3OF5BMl5BanBnXkFtZTgwNTU3NDA2MTI@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '11',
+   },
+   {
+     title: 'Pride and Prejudice and Zombies',
+     released_on: '2016',
+     genre: [
+       'Action',
+       'Horror',
+       'Romance',
+     ],
+     description: 'Five sisters in 19th century England must cope with the pressures to marry while protecting themselves from a growing population of zombies.',
+     actors: [
+       'Lily James',
+       'Sam Riley',
+       'Bella Heathcote',
+       'Ellie Bamber',
+       'Millie Brady',
+       'Suki Waterhouse',
+       'Douglas Booth',
+       'Sally Phillips',
+       'Charles Dance',
+       'Jack Huston',
+       'Lena Headey',
+       'Matt Smith',
+     ],
+     production_companies: [
+       'Cross Creek Pictures',
+       'MadRiver Pictures',
+       'QC Entertainment',
+       'Allison Shearmur Productions',
+       'Handsomecharlie Films',
+       'Head Gear Films',
+       'PPZ Holdings',
+     ],
+     run_time: '1 Hour 48 Minutes',
+     rating: 'PG-13',
+     director: 'Burr Steers',
+     writing_credits: [
+       'Burr Steers',
+       'Jane Austen',
+       'Seth Grahame-Smith',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjE1MzA3NzYxMl5BMl5BanBnXkFtZTgwMzQ0NDA5NzE@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '12',
+   },
+   {
+     title: 'Easy A',
+     released_on: '2010',
+     genre: [
+       'Comedy',
+       'Drama',
+       'Romance',
+     ],
+     description: 'A clean-cut high school student relies on the school\'s rumor mill to advance her social and financial standing.',
+     actors: [
+       'Emma Stone',
+       'Penn Badgley',
+       'Amanda Bynes',
+       'Dan Byrd',
+       'Thomas Haden Church',
+       'Patricia Clarkson',
+       'Stanley Tucci',
+       'Lisa Kudrow',
+       'Aly Michalka',
+     ],
+     production_companies: [
+       'Screen Gems',
+       'Olive Bridge Entertainment',
+     ],
+     run_time: '1 Hour 32 Minutes',
+     rating: 'PG-13',
+     director: 'Will Gluck',
+     writing_credits: 'Bert V. Royal',
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjE1MzU1Mjg1OF5BMl5BanBnXkFtZTcwNDc3NDQ2Mw@@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '13',
+   },
+   {
+     title: 'Fired Up!',
+     released_on: '2009',
+     genre: 'Comedy',
+     description: 'The two most popular guys in high school decide to ditch football camp for cheerleader camp. For the girls and for the glory.',
+     actors: [
+       'Nicholas D\'Agosto',
+       'Eric Christian Olsen',
+       'Sarah Roemer',
+       'Molly Sims',
+       'Danneel Ackles',
+       'David Walton',
+       'Adhir Kalyan',
+       'AnnaLynne McCord',
+       'Juliette Goglia',
+       'Philip Baker Hall',
+       'John Michael Higgins',
+       'Smith Cho',
+       'Margo Harshman',
+       'Hayley Marie Norman',
+       'Jake Sandvig',
+       'Nicole Tubiola',
+       'Collins Pennie',
+       'Edie McClurg',
+     ],
+     production_companies: [
+       'Screen Gems',
+       'Moving Pictures, DPI',
+       'Gross Entertainment',
+       'Weinstock Productions',
+     ],
+     run_time: '1 Hour 30 Minutes',
+     rating: 'PG-13',
+     director: 'Will Gluck',
+     writing_credits: 'Will Gluck',
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjE3MDMxMTQ0M15BMl5BanBnXkFtZTcwNTA0MDAyMg@@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '14',
+   },
+   {
+     title: 'Ant-Man',
+     released_on: '2015',
+     genre: [
+       'Action',
+       'Adventure',
+       'Comedy',
+     ],
+     description: 'Armed with a super-suit with the astonishing ability to shrink in scale but increase in strength, cat burglar Scott Lang must embrace his inner hero and help his mentor, Dr. Hank Pym, plan and pull off a heist that will save the world.',
+     actors: [
+       'Paul Rudd',
+       'Michael Douglas',
+       'Evangeline Lilly',
+       'Corey Stoll',
+       'Bobby Cannavale',
+       'Judy Greer',
+       'Michael Peña',
+       'Abby Ryder Fortson',
+     ],
+     production_companies: [
+       'Marvel Studios',
+       'Walt Disney Pictures',
+       'Gary Sanchez Productions',
+     ],
+     run_time: '1 Hour 57 Minutes',
+     rating: 'PG-13',
+     director: 'Peyton Reed',
+     writing_credits: [
+       'Edgar Wright',
+       'Joe Cornish',
+       'Adam McKay',
+       'Paul Rudd',
+       'Stan Lee',
+       'Larry Lieber',
+       'Jack Kirby',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjM2NTQ5Mzc2M15BMl5BanBnXkFtZTgwNTcxMDI2NTE@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '15',
+   },
+   {
+     title: 'Captain America: Civil War',
+     released_on: '2016',
+     genre: [
+       'Action',
+       'Adventure',
+       'Sci-Fi',
+     ],
+     description: 'Political interference in the Avengers\' activities causes a rift between former allies Captain America and Iron Man.',
+     actors: [
+       'Chris Evans',
+       'Robert Downey Jr.',
+       'Scarlett Johansson',
+       'Sebastian Stan',
+       'Anthony Mackie',
+       'Don Cheadle',
+       'Jeremy Renner',
+       'Chadwick Boseman',
+       'Paul Bettany',
+       'Elizabeth Olsen',
+       'Paul Rudd',
+       'Emily VanCamp',
+       'Tom Holland',
+       'Daniel Brühl',
+       'Frank Grillo',
+     ],
+     production_companies: [
+       'Marvel Studios',
+       'Vita-Ray Dutch Productions (III)',
+       'Studio Babelsberg',
+       'Marvel Entertainment',
+       'Walt Disney Pictures',
+       'Deluxe Digital Studios',
+     ],
+     run_time: '2 Hours 27 Minutes',
+     rating: 'PG-13',
+     director: [
+       'Anthony Russo',
+       'Joe Russo',
+     ],
+     writing_credits: [
+       'Christopher Markus',
+       'Stephen McFeely',
+       'Joe Simon',
+       'Jack Kirby',
+       'Mark Millar',
+       'Steven McNiven',
+     ],
+     thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMjQ0MTgyNjAxMV5BMl5BanBnXkFtZTgwNjUzMDkyODE@._V1_UX182_CR0,0,182,268_AL_.jpg',
+     id: '16',
+   },
+   {
+    title: 'Captain America: The First Avenger',
+    released_on: '2011',
+    genre: [
+      'Action',
+      'Adventure',
+      'Sci-Fi',
+    ],
+    description: 'Steve Rogers, a rejected military soldier transforms into Captain America after taking a dose of a "Super-Soldier serum." But being Captain America comes at a price as he attempts to take down a war monger and a terrorist organization.',
+    actors: [
+      'Chris Evans',
+      'Hayley Atwell',
+      'Sebastian Stan',
+      'Tommy Lee Jones',
+      'Hugo Weaving',
+      'Dominic Cooper',
+      'Richard Armitage',
+      'Stanley Tucci',
+      'Toby Jones',
+      'Neal McDonough',
+      'Derek Luke',
+      'Kenneth Choi',
+      'JJ Feild',
+      'Bruno Ricci',
+      'Lex Shrapnel',
+    ],
+    production_companies: [
+      'Paramount Pictures',
+      'Marvel Entertainment',
+      'Marvel Studios',
+    ],
+    run_time: '2 Hours 4 Minutes',
+    rating: 'PG-13',
+    director: 'Joe Johnston',
+    writing_credits: [
+      'Christopher Markus',
+      'Stephen McFeely',
+      'Joe Simon',
+      'Jack Kirby',
+    ],
+    thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMTYzOTc2NzU3N15BMl5BanBnXkFtZTcwNjY3MDE3NQ@@._V1_UX182_CR0,0,182,268_AL_.jpg',
+    id: '17',
+   },
+   {
+    title: 'Captain America: The Winter Soldier',
+    released_on: '2014',
+    genre: [
+      'Action',
+      'Adventure',
+      'Sci-Fi',
+    ],
+    description: 'As Steve Rogers struggles to embrace his role in the modern world, he teams up with a fellow Avenger and S.H.I.E.L.D agent, Black Widow, to battle a new threat from history: an assassin known as the Winter Soldier.',
+    actors: [
+      'Chris Evans',
+      'Samuel L. Jackson',
+      'Scarlett Johansson',
+      'Robert Redford',
+      'Sebastian Stan',
+      'Anthony Mackie',
+      'Cobie Smulders',
+      'Frank Grillo',
+      'Maximiliano Hernández',
+      'Emily VanCamp',
+      'Hayley Atwell',
+      'Toby Jones',
+    ],
+    production_companies: [
+      'Marvel Entertainment',
+      'Marvel Studios',
+      'Walt Disney Pictures',
+    ],
+    run_time: '2 Hours 16 Minutes',
+    rating: 'PG-13',
+    director: [
+      'Anthony Russo',
+      'Joe Russo',
+    ],
+    writing_credits: [
+      'Christopher Markus',
+      'Stephen McFeely',
+      'Joe Simon',
+      'Jack Kirby',
+    ],
+    thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMzA2NDkwODAwM15BMl5BanBnXkFtZTgwODk5MTgzMTE@._V1_UY268_CR1,0,182,268_AL_.jpg',
+    id: '18',
+   },
+   {
+    title: 'Dear Eleanor',
+    released_on: '2016',
+    genre: [
+      'Adventure',
+      'Comedy',
+      'Drama',
+    ],
+    description: 'Two teenage girls travel across the U.S. in 1962, during the chaos of the Cuban missile crisis, in search of Eleanor Roosevelt.',
+    actors: [
+      'Liana Liberato',
+      'Isabelle Fuhrman',
+      'Josh Lucas',
+      'Luke Wilson',
+      'Joel Courtney',
+      'Jessica Alba',
+    ],
+    production_companies: 'Nine Nights',
+    run_time: '1 Hour 29 Minutes',
+    rating: 'PG-13',
+    director: 'Kevin Connolly',
+    writing_credits: [
+      'Cecilia Contreras',
+      'Amy Garcia',
+    ],
+    thumbnail: 'https://images-na.ssl-images-amazon.com/images/M/MV5BMDMxNjAyNjctZDk2MS00MTFlLWJjNWItNWY3MmJmZTg4ZWQxXkEyXkFqcGdeQXVyMjA1MzUyODk@._V1_UY268_CR3,0,182,268_AL_.jpg',
+    id: '19',
+   },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+  //  {
+  //   title: '',
+  //   released_on: '',
+  //   genre: [],
+  //   description: '',
+  //   actors: [],
+  //   production_companies: [],
+  //   run_time: '',
+  //   rating: '',
+  //   director: '',
+  //   writing_credits: [],
+  //   thumbnail: '',
+  //  },
+
+   //little mermaid, beauty & the beast x 2, juno, a cinderella story,   marvel, star wars, house bunny,  penelope
+   //tv pride and prejudice, sailor moon, 7 deadly sins, durarara!, dr who, star trek, digimon, bones, castle, gg, jessica jones, alien show tbs, that 70's show, lizzie mcguire,
+ ];
