+++ src/app/movie-list/movie-list.component.html
@@ -1,0 +1,18 @@
+<input type="text"
+            placeholder="Search List"
+            class="filter-input"
+            [(ngModel)]="searchText">
+<div >
+  <ul>
+    <li *ngFor="let movie of movies | searchBy: ['title']: [searchText]">
+      <a[routerLink]="['/movie', movie.id]">
+        <h2>{{ movie.title }}</h2>
+      <!-- </a> -->
+      <img src="{{ movie.thumbnail }}" alt="">
+      <p>{{ movie.released_on }}</p>
+      <p>{{ movie.genre }}</p>
+      <p>{{ movie.production_companies }}</p>
+      <p>{{ movie.description }}</p>
+    </li>
+  </ul>
+</div>
