+++ src/app/movie-list/movie-list.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+
+import { MovieService } from '../services/movie.service';
+
+@Component({
+  templateUrl: './movie-list.component.html',
+  styleUrls: ['./movie-list.component.css']
+})
+export class MovieListComponent implements OnInit {
+  public movies;
+  public searchText: string;
+
+  constructor(private movieService: MovieService) { }
+
+  ngOnInit() {
+    this.movies = this.movieService.getMovies();
+  }
+
+}
