--- src/app/movies/movies.component.html
@@ -1,3 +1,0 @@
-<p>
-  movies works!
-</p>
