--- src/app/television/television.component.html
@@ -1,3 +1,0 @@
-<p>
-  television works!
-</p>
